🕯️ creating the retiever class to encansulate the get the last embeddings for users and jobs
