+++ src/app/services/mentor.py
@@ -1,0 +1,8 @@
+""" modeule to generate the recomendations """
+
+class Mentor():
+    """
+    """
+    def __init__(self):
+        self.embedding_path()
+
